requirements.txt
Appium-Python-Client == 4.2.0
openai
anthropic
pyyaml
opencv-python
easyocr
requests
pyshine
dill